+++ NEWS.md
@@ -58,6 +58,9 @@ usethis has a more sophisticated understanding of the default branch and gains s
   
 * `use_lifecycle()` now imports `lifecycle::deprecated()` (#1419).
 
+* `use_pkgdown()` automatically uses bs4 if you have the devel version
+   of pkgdown installed.
+
 * `use_code_of_conduct()` now requires a `contact` argument to supply contact
   details for reporting CoC violations (#1269).